Revert "Separate OS-Dependent code from OS-Independent code"
