+++ .gitignore
@@ -1,6 +1,35 @@
-env/*
-.vscode/*
-.eggs/*
+
+.eggs/
+setup.py
+dist/scout.app/Contents/Resources/Scout
+dist/scout.app/Contents/Resources/lib/python3.9/config
 *.pyz
-build/*
-dist/*
+build/scout/xref-scout.html
+build/scout/warn-scout.txt
+build/scout/Tree-02.toc
+build/scout/Tree-01.toc
+build/scout/Tree-00.toc
+build/scout/PYZ-00.toc
+build/scout/PKG-00.toc
+build/scout/PKG-00.pkg
+build/scout/EXE-00.toc
+build/scout/BUNDLE-00.toc
+build/scout/base_library.zip
+build/scout/Analysis-00.toc
+dist/scout
+windows/build/scout/xref-scout.html
+windows/build/scout/warn-scout.txt
+windows/build/scout/Tree-02.toc
+windows/build/scout/Tree-01.toc
+windows/build/scout/Tree-00.toc
+windows/build/scout/scout.exe.manifest
+windows/build/scout/PYZ-00.toc
+windows/build/scout/PKG-00.toc
+windows/build/scout/PKG-00.pkg
+windows/build/scout/EXE-00.toc
+windows/build/scout/base_library.zip
+windows/build/scout/Analysis-00.toc
+build/scout/scout.exe.manifest
+build/scout/runw.exe.eiw5f1je
+windows/dist/scout.exe
+dist/
