--- linux/paths.py
@@ -1,12 +1,0 @@
-import getpass
-
-user = '~'
-desktop = user + '/Desktop/'
-documents = user + '/Documents/'
-
-fileLoc = documents + '/Scout/'
-ymldir = fileLoc + 'settings.yml"
-logo_loc = fileLoc + 'scout_logo.gif'
-
-repo_url = 'https://raw.githubusercontent.com/leifadev/scout/main/'
-logo_url = repo_url + 'linux/scout_logo.gif'
